Move ConfirmAsync into shared components directory
