+++ metaspace/webapp/src/modules/GroupProfile/EditGroupProfile.vue
@@ -62,7 +62,7 @@
   import EditGroupMembersList from './EditGroupMembersList.vue';
   import { UserRole } from '../../api/user';
   import { encodeParams } from '../../url';
-  import ConfirmAsync from './ConfirmAsync';
+  import ConfirmAsync from '../../components/ConfirmAsync';
   import reportError from '../../lib/reportError';
   import emailRegex from '../../lib/emailRegex';
